fix: only auto-disable timestamps if we’re not in a terminal
On a systemd system, user sessions run under systemd, so our
"is systemd" detection always (correctly!) returns true, and
logs are always timeless. Obviously that’s not desired, so
we add a workaround it by detecting if stderr is a terminal;
this is imperfect but should cover the common case.